+++ zarrs/src/array.rs
@@ -186,7 +186,7 @@ pub fn chunk_shape_to_array_shape(chunk_shape: &[std::num::NonZeroU64]) -> Array
 /// The [`store_chunk_subset`](Array::store_chunk_subset) and [`store_array_subset`](Array::store_array_subset) are less preferred because they may incur decoding overhead and require careful usage if executed in parallel (see below).
 ///
 /// #### Direct IO (Linux)
-/// If using Linux, enabling direct IO with the [`FilesystemStore`](crate::storage::store::FilesystemStore) may improve write performance.
+/// If using Linux, enabling direct IO with the [`FilesystemStore`](https://docs.rs/zarrs_filesystem/latest/zarrs_filesystem/struct.FilesystemStore.html) may improve write performance.
 ///
 /// Currently, the most performant path for uncompressed writing is to reuse page aligned buffers via [`store_encoded_chunk`](Array::store_encoded_chunk).
 /// See [`zarrs` GitHub issue #58](https://github.com/LDeakin/zarrs/pull/58) for a discussion on this method.
@@ -848,7 +848,8 @@ pub fn bytes_to_ndarray<T: bytemuck::Pod>(
 
 #[cfg(test)]
 mod tests {
-    use crate::storage::store::{FilesystemStore, MemoryStore};
+    use crate::storage::store::MemoryStore;
+    use zarrs_filesystem::FilesystemStore;
 
     use super::*;